+++ tests/Framework/TnefDecoderTest.php
@@ -28,6 +28,35 @@ class TnefDecoderTest extends TestCase
     }
 
     /**
+     * Test TNEF decoding (#10193)
+     */
+    public function test_decompress_10193()
+    {
+        $body = file_get_contents(TESTS_DIR . 'src/winmail-10193.tnef');
+        $tnef = new \rcube_tnef_decoder();
+        $result = $tnef->decompress($body);
+
+        $this->assertSame("..Re: Vyzva_Teplotka_Ordinace_perioperacni_pece\0.html", trim($result['message']['name']));
+        $this->assertSame('text', $result['message']['type']);
+        $this->assertSame('html', $result['message']['subtype']);
+        $this->assertSame(16664, $result['message']['size']);
+        $this->assertSame(16664, strlen($result['message']['stream']));
+        $this->assertCount(3, $result['attachments']);
+        $this->assertSame(34965, $result['attachments'][0]['size']);
+        $this->assertSame(34965, strlen($result['attachments'][0]['stream']));
+        $this->assertTrue(str_starts_with($result['attachments'][0]['stream'], "\x89PNG"));
+        $this->assertSame('73471187-19a4-4a80-ac0b-6e776ff823fc', $result['attachments'][0]['content-id']);
+        $this->assertSame(39201, $result['attachments'][1]['size']);
+        $this->assertSame(39201, strlen($result['attachments'][1]['stream']));
+        $this->assertTrue(str_starts_with($result['attachments'][1]['stream'], "\x89PNG"));
+        $this->assertSame('32c2ea33-0a82-4941-9843-9b5bbd7e9cb2', $result['attachments'][1]['content-id']);
+        $this->assertSame(37205, $result['attachments'][2]['size']);
+        $this->assertSame(37205, strlen($result['attachments'][2]['stream']));
+        $this->assertTrue(str_starts_with($result['attachments'][2]['stream'], "\x89PNG"));
+        $this->assertSame('2be4b893-2b78-4442-81c0-8ddac66487c5', $result['attachments'][2]['content-id']);
+    }
+
+    /**
      * Test TNEF decoding
      */
     public function test_decompress_body()